requirements.txt
biopython==1.79
ete3==3.1.2
numpy==1.21.2
pandas==1.3.3
tqdm==4.62.3
xmltodict==0.12.0
argparse==1.4.0
scikit-learn==1.2.2
